move some test helper funcs to helper file
